test(CH): move Market default to test module
